split client_tcert_pool.go into multiple files
