+++ openchain/crypto/client_tcert_pool.go
@@ -19,13 +19,6 @@ under the License.
 
 package crypto
 
-import (
-	"errors"
-	"fmt"
-	"sync"
-	"time"
-)
-
 type tCertPool interface {
 	init(client *clientImpl) error
 
@@ -36,276 +29,4 @@ type tCertPool interface {
 	GetNextTCert() (tCert, error)
 
 	AddTCert(tCert tCert) error
-}
-
-type tCertPoolMultithreadingImpl struct {
-	client *clientImpl
-
-	tCertChannel         chan tCert
-	tCertChannelFeedback chan struct{}
-	done                 chan struct{}
-}
-
-func (tCertPool *tCertPoolMultithreadingImpl) Start() (err error) {
-	// Start the filler
-	go tCertPool.filler()
-
-	return
-}
-
-func (tCertPool *tCertPoolMultithreadingImpl) Stop() (err error) {
-	// Stop the filler
-	tCertPool.done <- struct{}{}
-
-	// Store unused TCert
-	tCertPool.client.node.debug("Store unused TCerts...")
-
-	tCerts := []tCert{}
-	for {
-		if len(tCertPool.tCertChannel) > 0 {
-			tCerts = append(tCerts, <-tCertPool.tCertChannel)
-		} else {
-			break
-		}
-	}
-
-	tCertPool.client.node.debug("Found %d unused TCerts...", len(tCerts))
-
-	tCertPool.client.node.ks.storeUnusedTCerts(tCerts)
-
-	tCertPool.client.node.debug("Store unused TCerts...done!")
-
-	return
-}
-
-func (tCertPool *tCertPoolMultithreadingImpl) GetNextTCert() (tCert tCert, err error) {
-	for i := 0; i < 3; i++ {
-		tCertPool.client.node.debug("Getting next TCert... %d out of 3", i)
-		select {
-		case tCert = <-tCertPool.tCertChannel:
-			break
-		case <-time.After(30 * time.Second):
-			tCertPool.client.node.error("Failed getting a new TCert. Buffer is empty!")
-
-			//return nil, errors.New("Failed getting a new TCert. Buffer is empty!")
-		}
-		if tCert != nil {
-			// Send feedback to the filler
-			tCertPool.tCertChannelFeedback <- struct{}{}
-			break
-		}
-	}
-
-	if tCert == nil {
-		// TODO: change error here
-		return nil, errors.New("Failed getting a new TCert. Buffer is empty!")
-	}
-
-	tCertPool.client.node.debug("Cert [% x].", tCert.GetCertificate().Raw)
-
-	// Store the TCert permanently
-	tCertPool.client.node.ks.storeUsedTCert(tCert)
-
-	tCertPool.client.node.debug("Getting next TCert...done!")
-
-	return
-}
-
-func (tCertPool *tCertPoolMultithreadingImpl) AddTCert(tCert tCert) (err error) {
-	tCertPool.client.node.debug("New TCert added.")
-	tCertPool.tCertChannel <- tCert
-
-	return
-}
-
-func (tCertPool *tCertPoolMultithreadingImpl) init(client *clientImpl) (err error) {
-	tCertPool.client = client
-
-	tCertPool.tCertChannel = make(chan tCert, client.node.conf.getTCertBathSize()*2)
-	tCertPool.tCertChannelFeedback = make(chan struct{}, client.node.conf.getTCertBathSize()*2)
-	tCertPool.done = make(chan struct{})
-
-	return
-}
-
-func (tCertPool *tCertPoolMultithreadingImpl) filler() {
-	// Load unused TCerts
-	stop := false
-	full := false
-	for {
-		// Check if Stop was called
-		select {
-		case <-tCertPool.done:
-			tCertPool.client.node.debug("Force stop!")
-			stop = true
-		default:
-		}
-		if stop {
-			break
-		}
-
-		tCertDER, err := tCertPool.client.node.ks.loadUnusedTCert()
-		if err != nil {
-			tCertPool.client.node.error("Failed loading TCert: [%s]", err)
-			break
-		}
-		if tCertDER == nil {
-			tCertPool.client.node.debug("No more TCerts in cache!")
-			break
-		}
-
-		tCert, err := tCertPool.client.getTCertFromDER(tCertDER)
-		if err != nil {
-			tCertPool.client.node.error("Failed paring TCert [% x]: [%s]", tCertDER, err)
-
-			continue
-		}
-
-		// Try to send the tCert to the channel if not full
-		select {
-		case tCertPool.tCertChannel <- tCert:
-			tCertPool.client.node.debug("TCert send to the channel!")
-		default:
-			tCertPool.client.node.debug("Channell Full!")
-			full = true
-		}
-		if full {
-			break
-		}
-	}
-
-	tCertPool.client.node.debug("Load unused TCerts...done!")
-
-	if !stop {
-		ticker := time.NewTicker(1 * time.Second)
-		for {
-			select {
-			case <-tCertPool.done:
-				stop = true
-				tCertPool.client.node.debug("Done signal.")
-			case <-tCertPool.tCertChannelFeedback:
-				tCertPool.client.node.debug("Feedback received. Time to check for tcerts")
-			case <-ticker.C:
-				tCertPool.client.node.debug("Time elapsed. Time to check for tcerts")
-			}
-
-			if stop {
-				tCertPool.client.node.debug("Quitting filler...")
-				break
-			}
-
-			if len(tCertPool.tCertChannel) < tCertPool.client.node.conf.getTCertBathSize() {
-				tCertPool.client.node.debug("Refill TCert Pool. Current size [%d].",
-					len(tCertPool.tCertChannel),
-				)
-
-				var numTCerts int = cap(tCertPool.tCertChannel) - len(tCertPool.tCertChannel)
-				if len(tCertPool.tCertChannel) == 0 {
-					numTCerts = cap(tCertPool.tCertChannel) / 10
-				}
-
-				tCertPool.client.node.info("Refilling [%d] TCerts.", numTCerts)
-
-				err := tCertPool.client.getTCertsFromTCA(numTCerts)
-				if err != nil {
-					tCertPool.client.node.error("Failed getting TCerts from the TCA: [%s]", err)
-				}
-			}
-		}
-	}
-
-	tCertPool.client.node.debug("TCert filler stopped.")
-}
-
-type tCertPoolSingleThreadImpl struct {
-	client *clientImpl
-
-	len    int
-	tCerts []tCert
-	m      sync.Mutex
-}
-
-func (tCertPool *tCertPoolSingleThreadImpl) Start() (err error) {
-	tCertPool.m.Lock()
-	defer tCertPool.m.Unlock()
-
-	tCertPool.client.node.debug("Starting TCert Pool...")
-
-	// Load unused TCerts if any
-	tCertDERs, err := tCertPool.client.node.ks.loadUnusedTCerts()
-	if err != nil {
-		tCertPool.client.node.error("Failed loading TCerts from cache: [%s]", err)
-
-		return
-	}
-	if len(tCertDERs) == 0 {
-		tCertPool.client.node.debug("No more TCerts in cache! Load new from TCA.")
-
-		tCertPool.client.getTCertsFromTCA(tCertPool.client.node.conf.getTCertBathSize())
-	} else {
-		tCertPool.client.node.debug("TCerts in cache found! Loading them...")
-
-		for _, tCertDER := range tCertDERs {
-			tCert, err := tCertPool.client.getTCertFromDER(tCertDER)
-			if err != nil {
-				tCertPool.client.node.error("Failed paring TCert [% x]: [%s]", tCertDER, err)
-
-				continue
-			}
-			tCertPool.AddTCert(tCert)
-		}
-	}
-
-	return
-}
-
-func (tCertPool *tCertPoolSingleThreadImpl) Stop() (err error) {
-	tCertPool.m.Lock()
-	defer tCertPool.m.Unlock()
-
-	tCertPool.client.node.debug("Found %d unused TCerts...", tCertPool.len)
-
-	tCertPool.client.node.ks.storeUnusedTCerts(tCertPool.tCerts[:tCertPool.len])
-
-	tCertPool.client.node.debug("Store unused TCerts...done!")
-
-	return
-}
-
-func (tCertPool *tCertPoolSingleThreadImpl) GetNextTCert() (tCert tCert, err error) {
-	tCertPool.m.Lock()
-	defer tCertPool.m.Unlock()
-
-	if tCertPool.len <= 0 {
-		// Reload
-		if err := tCertPool.client.getTCertsFromTCA(tCertPool.client.node.conf.getTCertBathSize()); err != nil {
-
-			return nil, fmt.Errorf("Failed loading TCerts from TCA")
-		}
-	}
-
-	tCert = tCertPool.tCerts[tCertPool.len-1]
-	tCertPool.len--
-
-	return
-}
-
-func (tCertPool *tCertPoolSingleThreadImpl) AddTCert(tCert tCert) (err error) {
-	tCertPool.client.node.debug("Adding new Cert [% x].", tCert.GetCertificate().Raw)
-
-	tCertPool.len++
-	tCertPool.tCerts[tCertPool.len-1] = tCert
-
-	return nil
-}
-
-func (tCertPool *tCertPoolSingleThreadImpl) init(client *clientImpl) (err error) {
-	tCertPool.client = client
-
-	tCertPool.client.node.debug("Init TCert Pool...")
-
-	tCertPool.tCerts = make([]tCert, tCertPool.client.node.conf.getTCertBathSize())
-	tCertPool.len = 0
-
-	return
 }
